started with 2nd part tests

+++ crss/libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: ivmirand <ivmirand@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/17 12:09:27 by ivmirand          #+#    #+#             */
-/*   Updated: 2024/09/17 12:42:21 by ivmirand         ###   ########.fr       */
+/*   Updated: 2024/09/25 14:49:38 by ivmirand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 char	*ft_strtrim(char const *s1, char const *set)
 {
-	int		i;
-	int		start;
-	int		end;
+	size_t	i;
+	size_t	start;
+	size_t	end;
 	char	*str;
 
 	start = 0;
-	end = (int)ft_strlen(s1);
+	end = ft_strlen(s1);
 	while (ft_strchr(set, s1[start]) != NULL && start < end)
 		start++;
 	while (ft_strchr(set, s1[end]) != NULL && end > 0)
